separate error type from handler
