+++ aws-datastore/src/main/java/com/amplifyframework/datastore/DataStoreErrorHandler.java
@@ -16,20 +16,16 @@
 package com.amplifyframework.datastore;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 
-import com.amplifyframework.api.graphql.GraphQLResponse;
 import com.amplifyframework.core.Amplify;
 import com.amplifyframework.core.Consumer;
 import com.amplifyframework.core.model.Model;
 import com.amplifyframework.logging.Logger;
 
-import java.util.Objects;
-
 /**
- * Just a ~type-alias for a consumer of DataStoreException.
+ * Just a type-alias for a consumer of DataStoreException.
  */
-public interface DataStoreErrorHandler extends Consumer<DataStoreErrorHandler.SyncError<? extends Model>> {
+public interface DataStoreErrorHandler extends Consumer<DataStoreError<? extends Model>> {
     /**
      * Gets a new instance of the {@link DefaultDataStoreErrorHandler}.
      * @return A new {@link DefaultDataStoreErrorHandler}
@@ -45,39 +41,8 @@ public interface DataStoreErrorHandler extends Consumer<DataStoreErrorHandler.Sy
         private DefaultDataStoreErrorHandler() {}
 
         @Override
-        public void accept(@NonNull SyncError<? extends Model> error) {
-            LOG.error("Error encountered in the DataStore:" + error.getError().getMessage());
-        }
-    }
-
-    final class SyncError<M extends Model> {
-        private final GraphQLResponse.Error error;
-        private final M local;
-        private final M remote;
-
-        public SyncError(
-                @NonNull GraphQLResponse.Error error,
-                @Nullable M local,
-                @Nullable M remote
-        ) {
-            this.error = Objects.requireNonNull(error);
-            this.local = local;
-            this.remote = remote;
-        }
-
-        @NonNull
-        public GraphQLResponse.Error getError() {
-            return error;
-        }
-
-        @Nullable
-        public M getLocal() {
-            return local;
-        }
-
-        @Nullable
-        public M getRemote() {
-            return remote;
+        public void accept(@NonNull DataStoreError<? extends Model> error) {
+            LOG.error("Error encountered in the DataStore: " + error.getError().getMessage());
         }
     }
 }
